minor refactoring removed this, reordered methods, props, removed vs hosting process

+++ branches/refactoring/LogExpert/Functionality/Log/LogBufferCacheEntry.cs
@@ -14,7 +14,7 @@ namespace LogExpert
 
 		public void Touch()
 		{
-			LastUseTimeStamp = (long)(Environment.TickCount & Int32.MaxValue);
+			LastUseTimeStamp = (Environment.TickCount & Int32.MaxValue);
 		}
 
 		internal LogBuffer LogBuffer { get; set; }
